[SofaImGui] Refactor methods
