+++ SofaImGui/CMakeLists.txt
@@ -84,6 +84,10 @@ set(HEADER_FILES
     ${SOFAIMGUI_SOURCE_DIR}/ImGuiGUI.h
     ${SOFAIMGUI_SOURCE_DIR}/ImGuiGUIEngine.h
     ${SOFAIMGUI_SOURCE_DIR}/ObjectColor.h
+        ${SOFAIMGUI_SOURCE_DIR}/windows/Performances.h
+        ${SOFAIMGUI_SOURCE_DIR}/windows/Log.h
+        ${SOFAIMGUI_SOURCE_DIR}/windows/Profiler.h
+
 )
 
 set(SOURCE_FILES
@@ -92,6 +96,11 @@ set(SOURCE_FILES
     ${SOFAIMGUI_SOURCE_DIR}/ImGuiGUIEngine.cpp
     ${SOFAIMGUI_SOURCE_DIR}/ObjectColor.cpp
     ${SOFAIMGUI_SOURCE_DIR}/initSofaImGui.cpp
+        ${SOFAIMGUI_SOURCE_DIR}/windows/Performances.cpp
+        ${SOFAIMGUI_SOURCE_DIR}/windows/Log.cpp
+        ${SOFAIMGUI_SOURCE_DIR}/windows/Profiler.cpp
+
+
 )
